Moved tests to correct package
