+++ src/test/java/frc/storage/commands/IndexBallsTest.java
@@ -1,9 +1,11 @@
-package frc.storage;
+package frc.storage.commands;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import frc.storage.StorageOutput;
+import frc.storage.StorageTestSensor;
 import frc.storage.command.IndexBalls;
