+++ src/test/java/frc/storage/commands/ShootBallTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.*;
 import org.junit.Test;
 
 import frc.storage.StorageOutput;
+import frc.storage.StorageTestSensor;
 import frc.storage.command.ShootBall;
 
 public class ShootBallTest {
